+++ l10n_it_ade/pyxbgen.py
@@ -376,8 +376,8 @@ def main(args):
 
 
 if __name__ == '__main__':
-    """pyxbgeb.py filename [schema] [fmlist]
-    fmlist is -u file -m module"""
+    # pyxbgeb.py filename [schema] [fmlist]
+    # fmlist is -u file -m module
     args = sys.argv[1:]
     while len(args) < 3:
         args.append('')