+++ l10n_it_einvoice_base/models/__init__.py
@@ -7,7 +7,6 @@
 from . import account_invoice
 from . import res_company
 from . import ir_attachment
-from . import res_partner
 from . import fetchmail_server
 from . import ir_mail_server
 from . import italy_ade_sender