+++ base_multireport/egg-info/history.rst
@@ -2,6 +2,8 @@
 ~~~~~~~~~~~~~~~~~~~~~~~
 
 * [FIX] Error "description_2_print() takes exactly 2 arguments (1 given)"
+* [FIX] Does not print fiscalcode on custom invoice
+
 
 10.0.0.2.6 (2019-05-20)
 ~~~~~~~~~~~~~~~~~~~~~~~